Move Zarr tests to test_zarr.py. Add a test for metadata with trailing whitespace.
